LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointermove-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
PointerMove

Test Description: This test checks if pointermove event triggers. Move your mouse over the black rectangle or slide it if you are using touchscreen.

Pointer Events pointermove Tests

The following pointer types were detected: (none).

Refresh the page to run the tests again with a different pointer type.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointermove event received
